requirements.txt
timm 
fairscale
torch
transformers
numpy
ruamel_yaml
opencv-python
matplotlib
cvxpy
scikit-learn
git+https://github.com/openai/CLIP.git
Pillow
pandas
dask

